requirements.txt
clouddrive==0.0.12.3
coloredlogs==15.0.1
Flask==3.1.0
# keyvalue_sqlite==1.0.7
# redis==5.2.0
PlexAPI==4.16.0
Requests==2.32.3
ruamel.yaml==0.18.6
gevent==24.11.1
marshmallow==3.23.1
pypinyin==0.53.0
Flask-Redis==0.4.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-APScheduler==1.13.1
flask-sqlalchemy==3.1.1
Celery==5.4.0
# Flask-Celery-Helper==1.1.0
# Flask-Execute==0.1.6
supervisor==4.2.5